Implement duplicate nodes async task.
